backend/requirements.txt
django==4.2.6
djangorestframework==3.14.0
pre-commit==3.4.0
